fix: wrap useSearchParams in Suspense boundary on /orgs/new
Split page into a server component (page.tsx) and client component
(new-org-client.tsx). The Suspense boundary satisfies Next.js's
requirement for useSearchParams and allows the route to build without
a prerender error.
